py-app/requirements.txt
# Python 3.14+
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=0.2.0
boto3>=1.34.84
python-dotenv>=1.0.1
